added userId to thought schema, added delete thoughts for user that was deleted, debugged routes in thought-routes

+++ routes/api/thought.routes.js
@@ -37,7 +37,7 @@ router.post("/", async (req, res) => {
         const addToUser = await User.findByIdAndUpdate(userId,
             { $push: { thoughts: newThought._id }},
             { new: true })
-        res.json({ result })
+            res.status(200).json({ message: 'Thought added successfully', thought: newThought })
     } catch (error) {
         res.status(500).json({ error: error.message });
     }
@@ -59,11 +59,20 @@ router.put('/:id', async (req, res) => {
 //Delete thought by id
 router.delete('/:id', async (req, res) => {
     try {
-        const result = await Thought.findByIdAndDelete(req.params.id)
-        if (!result) {
+        const thought = await Thought.findById(req.params.id)
+        if (!thought) {
+            return res.status(404).json({ message: 'Thought not found' });
+        }
+        const thoughtUserId = thought.userId
+        const removeThoughtFromUser = await User.findByIdAndUpdate(thoughtUserId,
+            {$pull: {thoughts: req.params.id }},
+            { new: true })
+        const deletedThought = await Thought.findByIdAndDelete(req.params.id)
+        // const deleteThoughtUserArr = await
+        if (!deletedThought) {
             return res.status(404).json({ message: 'Thought not found' })
         }
-        res.status(200).json({ message: 'Thought deleted successfully', thought: result })
+        res.status(200).json({ message: 'Thought deleted successfully', thought: deletedThought, user: removeThoughtFromUser })
     } catch (error) {
         res.status(500).json({ error: error.message });
     }
@@ -73,7 +82,7 @@ router.delete('/:id', async (req, res) => {
 router.post('/:thoughtId/reactions', async (req, res) => {
     try {
         const { thoughtId }= req.params
-        const { addReaction } = req.body
+        const addReaction = req.body 
         const result = await Thought.findByIdAndUpdate(thoughtId,
             { $push: { reactions: addReaction }},
             { new: true })
